+++ source/MView.Audio/Visualization/Controls/PolygonWaveFormControl.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="MView.Utilities.Audio.Controls.PolylineWaveFormControl"
+﻿<UserControl x:Class="MView.Audio.Visualization.Controls.PolygonWaveFormControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:MView.Utilities.Audio.Controls"
+             xmlns:local="clr-namespace:MView.Audio.Visualization.Controls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800" Foreground="{DynamicResource PrimaryBrush}">
     <Canvas x:Name="mainCanvas">
