+++ source/MView.Audio/Visualization/Controls/PolygonWaveFormControl.xaml.cs
@@ -13,31 +13,28 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace MView.Utilities.Audio.Controls
+namespace MView.Audio.Visualization.Controls
 {
     /// <summary>
-    /// Interaction logic for PolylineWaveFormControl.xaml
+    /// Interaction logic for PolygonWaveFormControl.xaml
     /// </summary>
-    public partial class PolylineWaveFormControl : UserControl, IWaveFormRenderer
+    public partial class PolygonWaveFormControl : UserControl, IWaveFormRenderer
     {
         int renderPosition;
         double yTranslate = 40;
         double yScale = 40;
+        double xScale = 2;
         int blankZone = 10;
 
-        readonly Polyline topLine = new Polyline();
-        readonly Polyline bottomLine = new Polyline();
+        readonly Polygon waveForm = new Polygon();
 
-        public PolylineWaveFormControl()
+        public PolygonWaveFormControl()
         {
             SizeChanged += OnSizeChanged;
             InitializeComponent();
-            topLine.Stroke = Foreground;
-            bottomLine.Stroke = Foreground;
-            topLine.StrokeThickness = 1;
-            bottomLine.StrokeThickness = 1;
-            mainCanvas.Children.Add(topLine);
-            mainCanvas.Children.Add(bottomLine);
+            waveForm.Stroke = Foreground;
+            waveForm.StrokeThickness = 1;
+            mainCanvas.Children.Add(waveForm);
         }
 
         void OnSizeChanged(object sender, SizeChangedEventArgs e)
@@ -52,29 +49,38 @@ namespace MView.Utilities.Audio.Controls
 
         private void ClearAllPoints()
         {
-            topLine.Points.Clear();
-            bottomLine.Points.Clear();
+            waveForm.Points.Clear();
+        }
+
+        private int Points
+        {
+            get { return waveForm.Points.Count / 2; }
         }
 
         public void AddValue(float maxValue, float minValue)
         {
-            int pixelWidth = (int)ActualWidth;
-            if (pixelWidth > 0)
+            int visiblePixels = (int)(ActualWidth / xScale);
+            if (visiblePixels > 0)
             {
                 CreatePoint(maxValue, minValue);
 
-                if (renderPosition > ActualWidth)
+                if (renderPosition > visiblePixels)
                 {
                     renderPosition = 0;
                 }
-                int erasePosition = (renderPosition + blankZone) % pixelWidth;
-                if (erasePosition < topLine.Points.Count)
+                int erasePosition = (renderPosition + blankZone) % visiblePixels;
+                if (erasePosition < Points)
                 {
                     double yPos = SampleToYPosition(0);
-                    topLine.Points[erasePosition] = new Point(erasePosition, yPos);
-                    bottomLine.Points[erasePosition] = new Point(erasePosition, yPos);
+                    waveForm.Points[erasePosition] = new Point(erasePosition * xScale, yPos);
+                    waveForm.Points[BottomPointIndex(erasePosition)] = new Point(erasePosition * xScale, yPos);
                 }
             }
+        }
+
+        private int BottomPointIndex(int position)
+        {
+            return waveForm.Points.Count - position - 1;
         }
 
         private double SampleToYPosition(float value)
@@ -84,17 +90,19 @@ namespace MView.Utilities.Audio.Controls
 
         private void CreatePoint(float topValue, float bottomValue)
         {
-            double topLinePos = SampleToYPosition(topValue);
-            double bottomLinePos = SampleToYPosition(bottomValue);
-            if (renderPosition >= topLine.Points.Count)
+            double topYPos = SampleToYPosition(topValue);
+            double bottomYPos = SampleToYPosition(bottomValue);
+            double xPos = renderPosition * xScale;
+            if (renderPosition >= Points)
             {
-                topLine.Points.Add(new Point(renderPosition, topLinePos));
-                bottomLine.Points.Add(new Point(renderPosition, bottomLinePos));
+                int insertPos = Points;
+                waveForm.Points.Insert(insertPos, new Point(xPos, topYPos));
+                waveForm.Points.Insert(insertPos + 1, new Point(xPos, bottomYPos));
             }
             else
             {
-                topLine.Points[renderPosition] = new Point(renderPosition, topLinePos);
-                bottomLine.Points[renderPosition] = new Point(renderPosition, bottomLinePos);
+                waveForm.Points[renderPosition] = new Point(xPos, topYPos);
+                waveForm.Points[BottomPointIndex(renderPosition)] = new Point(xPos, bottomYPos);
             }
             renderPosition++;
         }
