+++ source/MView.Audio/Visualization/Controls/PolylineWaveFormControl.xaml
@@ -1,12 +1,12 @@
-﻿<UserControl x:Class="MView.Utilities.Audio.Controls.SpectrumAnalyzerControl"
+﻿<UserControl x:Class="MView.Audio.Visualization.Controls.PolylineWaveFormControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:MView.Utilities.Audio.Controls"
+             xmlns:local="clr-namespace:MView.Audio.Visualization.Controls"
              mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-    <Canvas>
+             d:DesignHeight="450" d:DesignWidth="800" Foreground="{DynamicResource PrimaryBrush}">
+    <Canvas x:Name="mainCanvas">
         <Canvas.Background>
             <DrawingBrush TileMode="Tile" Viewport="0, 0, 10, 10" ViewportUnits="Absolute">
                 <DrawingBrush.Drawing>
@@ -21,6 +21,5 @@
                 </DrawingBrush.Drawing>
             </DrawingBrush>
         </Canvas.Background>
-        <Polyline x:Name="polyline" Stroke="{DynamicResource PrimaryBrush}" StrokeThickness="1"/>
     </Canvas>
 </UserControl>
