+++ source/MView.Audio/Visualization/IVisualizationPlugin.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MView.Utilities.Audio
+namespace MView.Audio.Visualization
 {
     public interface IVisualizationPlugin
     {
