+++ source/MView.sln
@@ -15,11 +15,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MView.Core", "MView.Core\MV
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MView.Plugin", "MView.Plugin\MView.Plugin.csproj", "{6C83E87B-EC16-4159-96B9-B0E30BCFD926}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MView.Test", "MView.Test\MView.Test.csproj", "{1B240A6B-EE32-41DF-B827-38D81820B234}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MView.Test", "MView.Test\MView.Test.csproj", "{1B240A6B-EE32-41DF-B827-38D81820B234}"
 	ProjectSection(ProjectDependencies) = postProject
 		{6C83E87B-EC16-4159-96B9-B0E30BCFD926} = {6C83E87B-EC16-4159-96B9-B0E30BCFD926}
 		{D63169F9-E644-4908-BDF6-F7EA42FE97FF} = {D63169F9-E644-4908-BDF6-F7EA42FE97FF}
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MView.Audio", "MView.Audio\MView.Audio.csproj", "{BEF4408D-47CF-403C-8A90-124C00B00BC4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -104,6 +106,24 @@ Global
 		{1B240A6B-EE32-41DF-B827-38D81820B234}.Release-x86|x64.Build.0 = Release|Any CPU
 		{1B240A6B-EE32-41DF-B827-38D81820B234}.Release-x86|x86.ActiveCfg = Release|Any CPU
 		{1B240A6B-EE32-41DF-B827-38D81820B234}.Release-x86|x86.Build.0 = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Debug|x64.Build.0 = Debug|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Debug|x86.Build.0 = Debug|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release|x64.ActiveCfg = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release|x64.Build.0 = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release|x86.ActiveCfg = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release|x86.Build.0 = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release-x86|Any CPU.ActiveCfg = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release-x86|Any CPU.Build.0 = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release-x86|x64.ActiveCfg = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release-x86|x64.Build.0 = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release-x86|x86.ActiveCfg = Release|Any CPU
+		{BEF4408D-47CF-403C-8A90-124C00B00BC4}.Release-x86|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
